Basic principal achieved
New implementation, need to do further test and allow for dynamic imports/ check performance with larger scale. Currently doing GeomCollection.contains. Not sure how that changes performance etc and need to see how big the final dataset is with global. If there are multiple threads - maybe it shuold be on the api handling itself, not the point checking